+++ webviz_subsurface/_models/inplace_volumes_model.py
@@ -6,7 +6,10 @@ import numpy as np
 import pandas as pd
 from pandas.api.types import is_numeric_dtype
 
-from .ensemble_set_model import EnsembleSetModel
+from webviz_subsurface._utils.ensemble_table_provider_set_factory import (
+    create_csvfile_providerset_from_paths,
+)
+
 from .parameter_model import ParametersModel
 
 
@@ -329,26 +332,35 @@ def filter_df(dframe: pd.DataFrame, filters: dict) -> pd.DataFrame:
     return dframe
 
 
-def extract_volumes(
-    ensemble_set_model: EnsembleSetModel,
+def extract_volframe_from_tableprovider(
+    ensemble_paths: dict,
     volfolder: str,
     volfiles: Dict[str, Any],
+    drop_failed_realizations: bool = True,
 ) -> pd.DataFrame:
-    """Aggregates volumetric files from an FMU ensemble.
+    """
+    Aggregates volumetric files from an FMU ensemble
     Files must be stored on standardized csv format.
     """
+
     dfs = []
     for volname, files in volfiles.items():
-        if isinstance(files, list):
-            volframes = [
-                ensemble_set_model.load_csv(Path(volfolder) / volfile)
-                for volfile in files
-            ]
-            df = merge_csv_files(volframes)
-        elif isinstance(files, str):
-            df = ensemble_set_model.load_csv(Path(volfolder) / files)
-        else:
+        if isinstance(files, str):
+            files = [files]
+
+        if not isinstance(files, list):
             raise ValueError("Wrong format of volfile value argument!")
+
+        volframes = []
+        for volfile in files:
+            table_provider_set = create_csvfile_providerset_from_paths(
+                ensemble_paths, str(Path(volfolder) / volfile), drop_failed_realizations
+            )
+            ensembledf = table_provider_set.get_aggregated_dataframe()
+            volframes.append(ensembledf)
+
+        # merge csvfiles from same SOURCE if more than one
+        df = merge_csv_files(volframes) if len(volframes) > 1 else volframes[0]
         df["SOURCE"] = volname
         dfs.append(df)